moved functions to .py
