+++ alphabase/peptide/fragment.py
@@ -933,69 +933,3 @@ def join_left(
     joined_index[left_indices] =  joined_index
 
     return joined_index
-
-
-# %% ../../nbdev_nbs/peptide/fragment.ipynb 39
-def annotate_fragments_from_speclib(
-    speclib: SpecLibBase, 
-    fragment_speclib: SpecLibBase,
-    verbose = True,
-):
-    """Reannotate an SpecLibBase library with fragments from a different SpecLibBase.
-    
-    Parameters
-    ----------
-    speclib: alphabase.spectral_library.library_base.SpecLibBase
-        Spectral library which contains the precursors to be annotated. All fragments mz and fragment intensities will be removed.
-
-    fragment_speclib: alphabase.spectral_library.library_base.SpecLibBase
-        Spectral library which contains the donor precursors whose fragments should be used.
-
-    Returns
-    -------
-
-    alphabase.spectral_library.library_base.SpecLibBase
-        newly annotated spectral library
- 
-    """
-    if verbose:
-        num_precursor_left = len(speclib.precursor_df)
-        num_precursor_right = len(fragment_speclib.precursor_df)
-        num_fragments_right = len(fragment_speclib.fragment_mz_df) * len(fragment_speclib.fragment_mz_df.columns)
-        logging.info(f'Speclib with {num_precursor_left:,} precursors will be reannotated with speclib with {num_precursor_right:,} precursors and {num_fragments_right:,} fragments')
-
-    # reannotation is based on mod_seq_hash column
-    hash_column_name = 'mod_seq_hash'
-
-    # create hash columns if missing
-    if hash_column_name not in speclib.precursor_df.columns:
-        speclib.hash_precursor_df()
-
-    if fragment_speclib not in fragment_speclib.precursor_df.columns:
-        fragment_speclib.hash_precursor_df()
-
-    speclib_hash = speclib.precursor_df[hash_column_name].values
-    fragment_speclib_hash = fragment_speclib.precursor_df[hash_column_name].values
-
-    speclib_indices = join_left(speclib_hash, fragment_speclib_hash)
-
-    matched_mask = (speclib_indices >= 0)
-
-    if verbose:
-        matched_count = np.sum(matched_mask)
-        not_matched_count = np.sum(~matched_mask)
-    
-        logging.info(f'A total of {matched_count:,} precursors were succesfully annotated, {not_matched_count:,} precursors were not matched')
-
-
-    frag_start_idx = fragment_speclib.precursor_df['frag_start_idx'].values[speclib_indices]
-    frag_end_idx = fragment_speclib.precursor_df['frag_end_idx'].values[speclib_indices]
-    
-    speclib._precursor_df = speclib._precursor_df[matched_mask].copy()
-    speclib._precursor_df['frag_start_idx'] = frag_start_idx[matched_mask]
-    speclib._precursor_df['frag_end_idx'] = frag_end_idx[matched_mask]
-
-    speclib._fragment_mz_df = fragment_speclib._fragment_mz_df.copy()
-    speclib._fragment_intensity_df = fragment_speclib._fragment_intensity_df.copy()
-
-    return speclib
